DABOM-345 refactor: 월초 배치 Redis 키 생성 로직 공용화
